backend/requirements.txt
flask>=2.2.5
flask-cors==3.0.10
flask-jwt-extended==4.4.4
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
matplotlib==3.8.4
pymysql==1.1.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
python-dotenv==1.0.0
gunicorn==20.1.0
pytest==7.3.1
pyyaml==6.0
flask-restful==0.3.9
scipy==1.13.0 